+++ src/components/reservant/navigation/navItems/Threads/Threads.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React, { useContext, useEffect, useState } from 'react'
 import OutsideClickHandler from '../../../../reusableComponents/OutsideClickHandler'
 import CommentRoundedIcon from '@mui/icons-material/CommentRounded'
 import AddCommentRoundedIcon from '@mui/icons-material/AddCommentRounded'
@@ -10,15 +10,14 @@ import {
 } from '../../../../../services/types'
 import { Button, CircularProgress, List, ListItemButton } from '@mui/material'
 import CheckIcon from '@mui/icons-material/Check'
-import DefaultPhoto from '../../../../../assets/images/user.jpg'
 import ThreadPreview from './ThreadPreview'
 import SearchIcon from '@mui/icons-material/Search'
-import Thread from './Thread'
 import { FetchError } from '../../../../../services/Errors'
 import InfiniteScroll from 'react-infinite-scroll-component'
-import InactiveThread from './InactiveThread'
 import { useTranslation } from 'react-i18next'
 import FriendSelector from '../../../../reusableComponents/FriendSelector'
+import renderUserPhotos from '../../../../../utils/DisplayUserPhotos'
+import { ThreadContext } from '../../../../../contexts/ThreadContext'
 
 const Threads: React.FC = () => {
   const [isPressed, setIsPressed] = useState<boolean>(false)
@@ -26,11 +25,11 @@ const Threads: React.FC = () => {
   const [isCreatingThread, setIsCreatingThread] = useState<boolean>(false)
   const [threadTitle, setThreadTitle] = useState<string>('')
   const [threads, setThreads] = useState<ThreadType[]>([])
-  const [activeThreads, setActiveThreads] = useState<ThreadType[]>([])
-  const [inactiveThreads, setInactiveThreads] = useState<ThreadType[]>([])
   const [page, setPage] = useState<number>(0)
   const [hasMore, setHasMore] = useState<boolean>(true)
   const [friendsToAdd, setFriendsToAdd] = useState<UserType[]>([])
+
+  const { handleThreadOpen } = useContext(ThreadContext)
 
   const [t] = useTranslation('global')
 
@@ -114,118 +113,6 @@ const Threads: React.FC = () => {
     }
   }
 
-  const handleThreadOpen = (thread: ThreadType) => {
-    setIsPressed(false)
-
-    const isActive = activeThreads.some(
-      activeThread => activeThread.threadId === thread.threadId
-    )
-
-    if (isActive) return
-
-    const isInactive = inactiveThreads.some(
-      inactiveThread => inactiveThread.threadId === thread.threadId
-    )
-
-    if (isInactive) {
-      handleThreadMaximize(thread)
-    } else {
-      if (activeThreads.length >= 2) {
-        handleThreadMinimize(activeThreads[0])
-      }
-      setActiveThreads(prevThreads => [...prevThreads, thread])
-    }
-  }
-
-  const handleThreadClose = (thread: ThreadType) => {
-    const isActive = activeThreads.some(
-      activeThread => activeThread.threadId === thread.threadId
-    )
-
-    if (isActive)
-      setActiveThreads(prevThreads =>
-        prevThreads.filter(
-          activeThread => activeThread.threadId !== thread.threadId
-        )
-      )
-    else
-      setInactiveThreads(prevThreads =>
-        prevThreads.filter(
-          inactiveThread => inactiveThread.threadId !== thread.threadId
-        )
-      )
-  }
-
-  const handleThreadMinimize = (thread: ThreadType) => {
-    setActiveThreads(prevThreads =>
-      prevThreads.filter(
-        activeThread => activeThread.threadId !== thread.threadId
-      )
-    )
-
-    if (inactiveThreads.length >= 5) {
-      handleThreadClose(inactiveThreads[0])
-    }
-
-    setInactiveThreads(prevThreads => [...prevThreads, thread])
-  }
-
-  const handleThreadMaximize = (thread: ThreadType) => {
-    setInactiveThreads(prevThreads =>
-      prevThreads.filter(
-        inactiveThread => inactiveThread.threadId !== thread.threadId
-      )
-    )
-
-    if (activeThreads.length >= 2) {
-      const earliestActiveThread = activeThreads[0]
-
-      setActiveThreads(prevThreads =>
-        prevThreads.filter((_, index) => index !== 0)
-      )
-
-      setInactiveThreads(prevThreads => {
-        const updatedInactiveThreads = [...prevThreads, earliestActiveThread]
-
-        if (updatedInactiveThreads.length > 5) {
-          return updatedInactiveThreads.slice(1)
-        }
-
-        return updatedInactiveThreads
-      })
-    }
-
-    setActiveThreads(prevThreads => [...prevThreads, thread])
-  }
-
-  const renderUserPhotos = (thread: ThreadType) => {
-    return thread.participants.length > 1 ? (
-      <div className="relative h-10 w-10">
-        <div className="absolute right-0 top-0 z-[0] flex h-8 w-8 items-center justify-center rounded-full bg-white dark:bg-black">
-          <img
-            src={getImage(thread.participants[0].photo, DefaultPhoto)}
-            className="absolute h-7 w-7 rounded-full"
-          />
-        </div>
-        <div className="absolute bottom-0 left-0 z-[1] flex h-8 w-8 items-center justify-center rounded-full bg-white dark:bg-black">
-          <img
-            src={getImage(thread.participants[1].photo, DefaultPhoto)}
-            className="absolute h-7 w-7 rounded-full"
-          />
-        </div>
-      </div>
-    ) : (
-      <div className="flex h-10 w-10 items-center justify-center">
-        <div className="flex h-10 w-10 items-center justify-center rounded-full bg-white dark:bg-black">
-          <img
-            src={getImage(thread.participants[0].photo, DefaultPhoto)}
-            className="h-9 w-9 rounded-full"
-          />
-        </div>
-      </div>
-    )
-  }
-
   const renderThreadsContent = () => {
     if (isLoadingThreads) {
       return (
@@ -262,7 +149,10 @@ const Threads: React.FC = () => {
             <ListItemButton
               key={thread.threadId}
               className="w-full rounded-md p-2 dark:hover:bg-grey-5"
-              onClick={() => handleThreadOpen(thread)}
+              onClick={() => {
+                handleThreadOpen(thread)
+                pressHandler()
+              }}
             >
               {
                 <ThreadPreview
@@ -353,31 +243,6 @@ const Threads: React.FC = () => {
           </div>
         )}
       </OutsideClickHandler>
-      <div
-        className={`absolute bottom-0 right-[0.5rem] max-w-[620px] ${inactiveThreads.length > 0 && 'right-[4rem]'} z-[0] z-[1] flex h-[400px] flex-row-reverse gap-[15px] ${activeThreads.length === 0 ? 'invisible' : activeThreads.length >= 2 ? 'w-[620]' : 'W-[310]'}`}
-      >
-        {activeThreads.map(activeThread => (
-          <Thread
-            key={activeThread.threadId}
-            thread={activeThread}
-            handleThreadClose={handleThreadClose}
-            handleThreadMinimize={handleThreadMinimize}
-          />
-        ))}
-      </div>
-      <div
-        className={`absolute bottom-0 right-[0.5rem] z-[0] z-[1] flex w-14 flex-col-reverse gap-2 py-4 ${inactiveThreads.length === 0 && 'hidden'}`}
-      >
-        {inactiveThreads.map(inactiveThread => (
-          <InactiveThread
-            key={inactiveThread.threadId}
-            thread={inactiveThread}
-            handleThreadMaximize={handleThreadMaximize}
-            renderUserPhotos={renderUserPhotos}
-            handleThreadClose={handleThreadClose}
-          />
-        ))}
-      </div>
     </div>
   )
 }
